Refactor API to predictable model
