+++ api/rawdata/serializers.py
@@ -1,0 +1,10 @@
+from share import models
+
+from api.base import ShareSerializer
+
+
+class RawDatumSerializer(ShareSerializer):
+
+    class Meta:
+        model = models.RawDatum
+        fields = ('id', 'suid', 'datum', 'sha256', 'date_modified', 'date_created', 'logs')
