+++ api/sourceregistrations/serializers.py
@@ -1,9 +1,6 @@
-from django.contrib.auth.models import AnonymousUser
-
 from rest_framework_json_api import serializers
 
 from share import models
-from share.models import ProviderRegistration, CeleryTaskResult
 
 from api.base import ShareSerializer
 
@@ -14,7 +11,7 @@ class ProviderRegistrationSerializer(ShareSerializer):
     submitted_by = serializers.HiddenField(default=serializers.CurrentUserDefault())
 
     def get_status(self, obj):
-        return ProviderRegistration.STATUS[obj.status]
+        return models.ProviderRegistration.STATUS[obj.status]
 
     class Meta:
         model = models.ProviderRegistration
