+++ api/sourceregistrations/urls.py
@@ -1,0 +1,8 @@
+from rest_framework.routers import DefaultRouter
+
+from api.sourceregistrations import views
+
+
+router = DefaultRouter()
+router.register(r'sourceregistrations', views.ProviderRegistrationViewSet, base_name='sourceregistration')
+urlpatterns = router.urls
