+++ api/sourceregistrations/views.py
@@ -1,9 +1,11 @@
-from rest_framework import viewsets, status
-from rest_framework.response import Response
+from rest_framework import status
+from rest_framework import viewsets
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
 
-from api.serializers import ProviderRegistrationSerializer
 from share.models import ProviderRegistration
+
+from api.sourceregistrations.serializers import ProviderRegistrationSerializer
 
 
 class ProviderRegistrationViewSet(viewsets.ModelViewSet):
