+++ api/views/__init__.py
@@ -1,7 +1,3 @@
 from .workflow import *  # noqa
 from .share import *  # noqa
-from .elasticsearch import *  # noqa
 from .feeds import *  # noqa
-from .registration import *  # noqa
-from .schema import *  # noqa
-from .banner import *  # noqa
