+++ tests/api/test_user_endpoint.py
@@ -14,35 +14,41 @@ def post_body_share_user():
     })
 
 
+@pytest.fixture(params=[
+    '/api/v2/user/',
+    '/api/v2/users/',
+])
+def endpoint(request):
+    return request.param
+
+
 @pytest.mark.django_db
 class TestSourcesGet:
-    endpoint = '/api/v2/user/'
 
-    def test_logged_in(self, client, share_user):
-        resp = client.get(self.endpoint, HTTP_AUTHORIZATION=share_user.authorization())
+    def test_logged_in(self, endpoint, client, share_user):
+        resp = client.get(endpoint, HTTP_AUTHORIZATION=share_user.authorization())
         assert resp.status_code == 200
         assert resp.json()['meta']['pagination']['count'] == 1
 
-    def test_not_logged_in(self, client):
-        resp = client.get(self.endpoint)
+    def test_not_logged_in(self, endpoint, client):
+        resp = client.get(endpoint)
         assert resp.status_code == 200
         assert resp.json()['meta']['pagination']['count'] == 0
 
 
 @pytest.mark.django_db
 class TestSourcesPost:
-    endpoint = '/api/v2/user/'
 
-    def test_unauthorized_post(self, client, post_body_share_user):
+    def test_unauthorized_post(self, endpoint, client, post_body_share_user):
         assert client.post(
-            self.endpoint,
+            endpoint,
             post_body_share_user,
             content_type='application/vnd.api+json'
         ).status_code == 401
 
-    def test_authorized_post(self, client, share_user, post_body_share_user):
+    def test_authorized_post(self, endpoint, client, share_user, post_body_share_user):
         assert client.post(
-            self.endpoint,
+            endpoint,
             post_body_share_user,
             content_type='application/vnd.api+json',
             HTTP_AUTHORIZATION=share_user.authorization(),
