+++ tests/api/test_validator.py
@@ -345,7 +345,7 @@ class TestValidator:
 
         kwargs['HTTP_AUTHORIZATION'] = 'Bearer {}'.format(trusted_user.accesstoken_set.first())
 
-        with mock.patch('api.views.workflow.disambiguate') as mock_disambiguate:
+        with mock.patch('api.normalizeddata.views.disambiguate') as mock_disambiguate:
             mock_disambiguate.delay().id = '123'
             assert response == client.post('/api/v2/normalizeddata/', *args, **kwargs)
 
@@ -373,7 +373,7 @@ class TestValidator:
 
         kwargs['HTTP_AUTHORIZATION'] = 'Bearer {}'.format(robot_user.accesstoken_set.first())
 
-        with mock.patch('api.views.workflow.disambiguate') as mock_disambiguate:
+        with mock.patch('api.normalizeddata.views.disambiguate') as mock_disambiguate:
             mock_disambiguate.delay().id = '123'
             response = client.post('/api/v2/normalizeddata/', *args, **kwargs)
